+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * SPSA tune - 938 iters at 20.0+0.20 (24.36 +- 8.47)
 * SPSA tune - 3580 iters at 20.0+0.20 (16.55 +- 7.10)
 * SPSA tune - 1077 iters at 20.0+0.20 (4.06 +- 3.05)
+* The en-passant square is no longer set when there are no legal en-passant moves (4.57 +- 4.38)
 * Implement a proper thread pool
 * Add mate distance pruning
 * Stop returning mate scores from NMP